service_editor/requirements.txt
aioamqp==0.14.0
boto3==1.13.24
botocore==1.16.24
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
docutils==0.15.2
google-api-core==1.19.1
google-auth==1.16.1
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
jmespath==0.10.0
moviepy==1.0.3
numpy==1.18.5
pamqp==2.3.0
Pillow==7.1.2
proglog==0.1.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-ffmpeg==1.0.11
pytz==2020.1
pytube==10.4.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
sentry-sdk==0.14.4
six==1.15.0
tqdm==4.46.1
urllib3==1.25.9
Wand==0.6.1
websocket-client==0.57.0
websockets==8.1
